Minor: Changed chapternumber type from float32 to float64

+++ src/mangaStructs/chapter.go
@@ -8,7 +8,7 @@ type Chapter struct {
 	ID             string
 	Manga          Manga
 	Title          string
-	ChapterNumber  float32
+	ChapterNumber  float64
 	Cover          Cover
 	RelationsShips []ChapterRelationship
 }
